+++ Source/Engine/Globals.cs
@@ -1,0 +1,9 @@
+﻿namespace Masterwatch.Source.Engine
+{
+    public static class Globals
+    {
+        public static string versionNumber = "Version: 1.4.0";
+        internal static int ScreenWidth = 1280;
+        internal static int ScreenHeight = 780;
+    }
+}