+++ Source/States/State.cs
@@ -1,0 +1,50 @@
+﻿using GeonBit.UI;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Content;
+using Microsoft.Xna.Framework.Graphics;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection.Metadata;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Masterwatch.Source.States
+{
+    public abstract class State
+    {
+        #region Fields
+
+        protected ContentManager _content;
+
+        protected GraphicsDevice _graphicsDevice;
+
+        protected Game1 _game;
+
+        protected UserInterface _userInterface; // Added field for UserInterface
+
+        #endregion
+
+        #region Methods
+
+        public State(Game1 game, GraphicsDevice graphicsDevice, ContentManager content)
+        {
+            _game = game;
+
+            _graphicsDevice = graphicsDevice;
+            _content = content;
+
+            UserInterface.Initialize(content, BuiltinThemes.hd);
+            _userInterface = new UserInterface();
+
+        }
+
+        public abstract void Draw(GameTime gameTime, SpriteBatch spriteBatch);
+
+        public abstract void Update(GameTime gameTime);
+
+        public abstract void PostUpdate(GameTime gameTime);
+
+        #endregion
+    }
+}