+++ CHANGELOG.md
@@ -6,10 +6,31 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+
+ # [1.4.0] - 2023-06-16
+ 
+## Added
+* Achievements screen (Not working yet)
+* Made 5 log/wood types and icons:
+    * Oak
+    * Birch
+    * Walnut
+    * Willow
+    * Redwood
+
+### Changed
+* Rewrite most of the code
+
+### Fixed
+
+* States actually work now
+
+***
+
  # [1.3.0] - 2023-06-14
  
 ## Added
-* Screens! (named as pages in the code)
+* Screens! (named as states in the code)
     * Main menu, settings, game screen, etc.
 
 * Some icon work
@@ -17,7 +38,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     * House icon (will be used later)
     * Game title icon (very basic)
 
-* Placeholder background for the menu
+* Placeholder background
 ### Changed
 * Game cursor
 
@@ -29,8 +50,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  # [1.2.0] - 2023-06-13 - Engine work
 Working on the engine.
 ## Added
-* Game engine additions:
-    * Pages, Objects, PageManager, ObjectManager
+* Game engine class additions
+    
 
  ***